refactor: extract common helpers and fix imports
