+++ internal/utils/validation.go
@@ -50,3 +50,13 @@ func HasUncommittedChanges(ctx context.Context) bool {
 	}
 	return output != ""
 }
+
+// ContainsString checks if a string is present in a slice of strings
+func ContainsString(slice []string, item string) bool {
+	for _, s := range slice {
+		if s == item {
+			return true
+		}
+	}
+	return false
+}
